Fix Bug#1 and Bug#4 in Issue#13
Because I don't have a typescript compiler, the changes are made base on my local modified js file. Thus, further test is required.
The key change is line#27, while the other lines are just for optimization.
By the way, things seem to be worse for Bug#7. Latter commits may fix it.