Fix Bugs Reported in Issue#13 (#14)
* Fix Bug#1 and Bug#4 in Issue#13 

Because I don't have a typescript compiler, the changes are made base on my local modified js file. Thus, further test is required.
The key change is line#27, while the other lines are just for optimization.
By the way, things seem to be worse for Bug#7. Latter commits may fix it.

* Fix Bug#2 in Issue#13

Just one line#44 is modified.

* Fix Bug#3 in Issue#13

Just one line#26 is modified.
However, things seem to be worse for Bug#7. Latter commits may fix it.

* Fix Bug#5 and Bug#6 in Issue#13

The changes basically make the editor not always "eat" the Enter key event.

* Fix Bug#7 in Issue#13 (for Emoji Suggester)

1. Add a function checkForInputBlock( ) to check if the cursor is in a code/math block. The function can be reused for latter commits.
2. Add a condition determined by checkForInputBlock( ) function to the update( ) function, to fix Bug#7 for Emoji Suggester board.

* Fix Bug#7 in Issue#13 (for Imme. Emoji Replace)

Add a condition determined by checkForInputBlock( ) function to the replaceHandler, to fix Bug#7 for Immediate Emoji Replacer.

* Fix Bug#7 in Issus#13 (for Preview mode)

Add a condition determined by checkForInputBlock( ) function to the emojiReplace( ) function, to fix Bug#7 for Preview mode.

* Fix TypeScript compatibility

* complementary commit for PR#14

Add line#84 to fixed the "when you press Enter in the Suggester it will insert the currently selected Emoji and go to the next line (because of the Enter key press)" problem.

Co-authored-by: phibr0 <bronzel.phillip@gmail.com>